Fix dropdowns.
