--- webapp/src/components/profile/dropdown_selector_style.tsx
@@ -1,103 +1,0 @@
-import styled from 'styled-components';
-
-export const DropdownSelectorStyle = styled.div`
-    .profile-dropdown--active {
-        .PlaybookRunProfileButton {
-            .Profile {
-                background-color: var(--button-bg-08);
-            }
-        }
-    }
-
-    .profile-dropdown {
-        .playbook-run-user-select__container {
-            z-index: 50;
-            position: absolute;
-        }
-    }
-
-    .playbook-run-user-select__menu-list {
-        padding: 0 0 12px;
-        border: none;
-    }
-
-    .playbook-run-user-select__input {
-        color: var(--center-channel-color);
-    }
-
-    .playbook-run-user-select {
-        border-radius: 4px;
-        -webkit-overflow-scrolling: touch;
-        background-color: var(--center-channel-bg);
-        border: 1px solid var(--center-channel-color-16);
-        max-height: 100%;
-        max-width: 340px;
-        overflow: hidden;
-        box-shadow: 0 8px 24px rgba(0, 0, 0, 0.12);
-    }
-
-    .playbook-run-user-select__option--is-selected {
-        background-color: var(--center-channel-color-08);
-    }
-
-    .playbook-run-user-select__option--is-focused {
-        background-color: var(--center-channel-color-08);
-        color: inherit;
-    }
-
-    .playbook-run-user-select__control {
-        margin: 12px;
-        border: none;
-        background: transparent;
-        box-shadow: inset 0 0 0 1px var(--center-channel-color-20);
-
-        &:hover {
-            box-shadow: inset 0 0 0 1px var(--center-channel-color-40);
-        }
-
-        &--is-focused,
-        &--is-focused:hover {
-            box-shadow: inset 0 0 0 2px var(--button-bg);
-        }
-    }
-
-    .playbook-run-user-select__option {
-        &:active {
-            background-color: var(--center-channel-color-08);
-        }
-    }
-
-    .playbook-run-user-select__value-container {
-        overflow: visible;
-        padding-left: 3.2rem;
-
-        &:before {
-            position: absolute;
-            left: 0.8rem;
-            color: var(--center-channel-color-56);
-            content: '\f349';
-            font-family: 'compass-icons', mattermosticons;
-            font-style: normal;
-            font-weight: normal;
-            speak: none;
-            display: inline-block;
-            text-decoration: inherit;
-            width: 1em;
-            margin-right: .2em;
-            text-align: center;
-            font-variant: normal;
-            text-transform: none;
-            line-height: 1em;
-            margin-left: .2em;
-            font-size: 16px;
-            -webkit-font-smoothing: antialiased;
-            -moz-osx-font-smoothing: grayscale;
-        }
-
-        .icon {
-            position: absolute;
-            left: -2.2rem;
-            top: 3px;
-        }
-    }
-`;
