+++ webapp/src/components/profile/profile_selector.tsx
@@ -16,7 +16,7 @@ import Profile from 'src/components/profile/profile';
 import ProfileButton from 'src/components/profile/profile_button';
 import {useClientRect} from 'src/hooks';
 import {PlaybookRunFilterButton} from '../backstage/styles';
-import {DropdownSelectorStyle} from 'src/components/profile/dropdown_selector_style';
+import Dropdown from 'src/components/dropdown';
 
 export interface Option {
     value: string;
@@ -307,68 +307,6 @@ const selectStyles: StylesConfig<Option, boolean> = {
             color: 'unset',
         };
     },
-};
-
-// styled components
-interface DropdownProps {
-    children: JSX.Element;
-    isOpen: boolean;
-    showOnRight?: boolean;
-    moveUp?: number;
-    target: JSX.Element;
-    onClose: () => void;
-}
-
-const ProfileDropdown = styled.div`
-    position: relative;
-`;
-
-const Blanket = styled.div`
-    bottom: 0;
-    left: 0;
-    top: 0;
-    right: 0;
-    position: fixed;
-    z-index: 1;
-`;
-
-interface ChildContainerProps {
-    moveUp?: number;
-    showOnRight?: boolean;
-}
-
-const ChildContainer = styled.div<ChildContainerProps>`
-    margin: 4px 0 0;
-    min-width: 20rem;
-    top: ${(props) => 27 - (props.moveUp || 0)}px;
-    ${(props) => props.showOnRight && css`
-        right: -55px;
-    `}
-`;
-
-const Dropdown = ({children, isOpen, showOnRight, moveUp, target, onClose}: DropdownProps) => {
-    if (!isOpen) {
-        return target;
-    }
-
-    const classes = classNames('PlaybookRunFilter', 'profile-dropdown',
-        'PlaybookRunFilter--active', 'profile-dropdown--active');
-
-    return (
-        <DropdownSelectorStyle>
-            <ProfileDropdown className={classes}>
-                {target}
-                <ChildContainer
-                    className='playbook-run-user-select__container'
-                    moveUp={moveUp}
-                    showOnRight={showOnRight}
-                >
-                    {children}
-                </ChildContainer>
-                <Blanket onClick={onClose}/>
-            </ProfileDropdown>
-        </DropdownSelectorStyle>
-    );
 };
 
 const getFullName = (firstName: string, lastName: string): string => {
